+++ HotelApp.Application/DTOs/Floor/EditFloorDTO.cs
@@ -9,7 +9,7 @@ namespace HotelApp.Application.DTOs.Floor
     public class EditFloorDTO
     {
         public int Id { get; set; }
-        public int Number {  get; set; }
+        public Int16 Number {  get; set; }
         public bool IsActive { get; set; }
     }
 }
