+++ HotelApp.Application/DTOs/Rooms/GetRoomByIdDTO.cs
@@ -12,7 +12,7 @@ namespace HotelApp.Application.DTOs.Rooms
 		public int Id { get; set; }
 		public string RoomNumber { get; set; }
 		public string Description { get; set; }
-		public int Floor { get; set; }
+		public Int16 Floor { get; set; }
 		public decimal PricePerNight { get; set; }
 		public string TypeName { get; set; }
 		public int MaxNumOfAdults { get; set; }
